+++ employees/src/main/java/com/netit/employees/exception/EmployeeNotFoundException.java
@@ -2,7 +2,7 @@ package com.netit.employees.exception;
 
 public class EmployeeNotFoundException extends RuntimeException{
     
-    public EmployeeNotFoundException(Long id) {
+    public EmployeeNotFoundException(int id) {
         super("Could not found the employee with id " + id);
         
     }
